rentry-ai-microservice/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
psycopg2-binary
python-dotenv